requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-stubs==3.14.5
idna==3.6
pdfkit==1.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.8.0
urllib3==2.1.0
